identity: Remove duplicated authOptions (v2)
'gophercloud.AuthOptions' is intended a single, flat "store" for
authentication-related options and maps to the flat stores of
environment variables (e.g. the typical 'OS_PASSWORD' and 'OS_USERNAME'
combo) or the 'auth' section of cloud entry in a 'clouds.yaml' file.
However, the request body for 'POST /tokens' is not (and never has been)
flat so this flat object needs to be converted into a proper non-flat
request body.

Normally this is achieved via use of 'CreateOptsBuilder' interfaces with
a single method signature, 'ToFooCreateMap(map[string]any, error)'
(where 'Foo' is the resource), that converts the non-flat 'CreateOpts'
struct to a non-flat JSON body For presumably historical reasons though,
we were not following this pattern for the 'v2/tokens' and 'v3/tokens'
modules. In place of a 'CreateOpts' struct we had an 'AuthOptionsV2'
struct. And instead of implementing the 'ToTokenCreateMap' method on a
'CreateOpts' struct, we were implementing a 'ToTokenV2CreateMap' on yet
*another* struct, 'AuthOptions', which mostly mimicked
'gophercloud.AuthOptions' but focused only on version-specific fields.

There's no need for this to be such a special creature. Restore some
sanity by renaming 'AuthOptionsV2' to 'CreateOpts' and adding a new,
minimal 'ToTokenCreateMap' method on same. Since we do still need some
mechanism to translate the 'AuthOptions' objects to these 'CreateOpts'
helpers, we also add a new helper, 'FromAuthOptions', which will convert
a 'gophercloud.AuthOptions' object to this new 'CreateOpts' object,
allowing us to drop the existing 'AuthOptions.ToTokenV2CreateMap'
helper.

This commit only tackles the v2 variant of the tokens module. A future
change will tackle the v3 variant, as it's significantly more involved.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>