identity: Move OAuth Auth header generation to common module
This code is rather generic OAuth 1a 'Authorization' header generation
code that we're soon going to need to be able to call from multiple
locations. Place it in a more generic location to allow that to happen.

We remove some of the configurability of this header in the process
since YAGNI.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>