identity: Remove duplicated authOptions (v3)
We previously tackled this for the v2 identity API. Now do it for the v3
API. This one is more involved since there's just a lot more logic
needed but the general flow is the exact same. Moving this allows us to
drop all of the request body generation code for the AuthOptions object,
massively simplifying the AuthOptsBuilder interface. It also eases the
path to merging the 'oauth1' authentication method into the main
'tokens' code.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>